+++ docs/source/other/release-notes.md
@@ -5,6 +5,9 @@
 **Internal changes & bugfixes**
 - General
 	- Fixed bug that incorrectly removed submodel reference from AAS when updating a submodel via PUT /submodels/{submodelId}
+- Endpoint
+	- HTTP
+		- the URL prefix /api/v3.x is now optional
 
 ## 1.3.0